Add no vowel doubling exception checks
