+++ packages/yoastseo/src/morphology/morphoHelpers/stemHelpers.js
@@ -18,3 +18,4 @@ const removeSuffixFromFullForm = function( exceptions, suffix, word ) {
 export {
 	removeSuffixFromFullForm,
 };
+
